backend/requirements.txt
Flask==2.3.2
Flask-Cors==4.0.0
Flask-RESTful==0.3.10
python-dotenv==1.0.0
Werkzeug==2.3.6
ultralytics==8.3.72
requests
flask-socketio
eventlet
opencv-python==4.8.0.76
Pillow==10.0.0
numpy==1.26.4
flask_sqlalchemy
pyjwt
